+++ src/ALC/ApplicationInsights.cs
@@ -1,62 +1,62 @@
-﻿using System;
-using Microsoft.ApplicationInsights.Extensibility;
-using Microsoft.ApplicationInsights;
+﻿using Microsoft.ApplicationInsights;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace PnP.PowerShell.ALC
 {
-    // implement telemtryclient as a singleton to avoid memory leaks
-    public static class TelemetryClientFactory
-  {
-    private static TelemetryClient _telemetryClient;
- 
-    public static TelemetryClient GetTelemetryClient()
-    {
-            
-      if (_telemetryClient == null)
-      {
-        
-            TelemetryConfiguration config = TelemetryConfiguration.CreateDefault();
-            _telemetryClient = new TelemetryClient(config);
-            config.InstrumentationKey = "a301024a-9e21-4273-aca5-18d0ef5d80fb";
-            //config..Context.Session.Id = Guid.NewGuid().ToString();
-            _telemetryClient.Context.Cloud.RoleInstance = "PnPPowerShell";
-            _telemetryClient.Context.Device.OperatingSystem = Environment.OSVersion.ToString();
-
-            
-
-      }
- 
-      return _telemetryClient;
-    }
-  }
-
+    /// <summary>
+    /// Azure Application Insights Telemetry Client to allow for usage tracking
+    /// </summary>
     public class ApplicationInsights
     {
-        private TelemetryClient telemetryClient;
-        private static  Dictionary<string, string> telemetryProperties;
+        /// <summary>
+        /// Instance of the TelemetryClient to use to report back statistical data
+        /// </summary>
+        private TelemetryClient _telemetryClient;
+
+        /// <summary>
+        /// Information to provide with each telemetry log entry
+        /// </summary>
+        private static Dictionary<string, string> telemetryProperties;
         
-
-
+        /// <summary>
+        /// Initializes a new instance of the ApplicationInsights telemetry
+        /// </summary>
+        /// <param name="serverLibraryVersion">Version of SharePoint Online to which a connection has been established</param>
+        /// <param name="serverVersion">Version of the server to which a connection has been established</param>
+        /// <param name="initializationType">Information on what method has been used to establish a connection</param>
+        /// <param name="assemblyVersion">The PnP PowerShell version in use</param>
+        /// <param name="operatingSystem">The operating system on which PnP PowerShell is being used</param>
         public void Initialize(string serverLibraryVersion, string serverVersion, string initializationType, string assemblyVersion, string operatingSystem)
         {
-            telemetryClient = TelemetryClientFactory.GetTelemetryClient();
-            if (telemetryProperties == null ) {
-                telemetryProperties = new Dictionary<string, string>
+            // Retrieve an instance of the telemetry client to use
+            _telemetryClient = TelemetryClientFactory.GetTelemetryClient();
+
+            // Define the base set of information to log with each trackable event in Azure Application Insights
+            if (telemetryProperties == null)
             {
-                { "ServerLibraryVersion", serverLibraryVersion },
-                { "ServerVersion", serverVersion },
-                { "ConnectionMethod", initializationType.ToString() },
-                { "Version", assemblyVersion },
-                { "Platform", "SPO" },
-                { "OperatingSystem", operatingSystem}
-            };
+                telemetryProperties = new Dictionary<string, string>
+                {
+                    { "ServerLibraryVersion", serverLibraryVersion },           // Version of SharePoint Online to which a connection has been established
+                    { "ServerVersion", serverVersion },                         // Version of the server to which a connection has been established
+                    { "ConnectionMethod", initializationType.ToString() },      // Information on what method has been used to establish a connection
+                    { "Version", assemblyVersion },                             // The PnP PowerShell version in use
+                    { "Platform", "SPO" },                                      // Platform to which the connection has been made
+                    { "OperatingSystem", operatingSystem}                       // The operating system on which PnP PowerShell is being used
+                };
             }
         }
+
+        /// <summary>
+        /// Sends information to Azure Application Insights to track an event
+        /// </summary>
+        /// <param name="cmdletName">Name of the PnP PowerShell cmdlet that is being executed</param>
+        /// <param name="properties">Additional information on the cmdlet being executed. Optional.</param>
         public void TrackEvent(string cmdletName, Dictionary<string, string> properties = null)
         {
+            // Take the base set of information to log
             var localProps = telemetryProperties;
+
+            // If additional properties have been provided to log, add them to the base set of information to log
             if (properties != null)
             {
                 foreach (var prop in properties)
@@ -64,7 +64,9 @@ namespace PnP.PowerShell.ALC
                     localProps.Add(prop.Key, prop.Value);
                 }
             }
-            telemetryClient.TrackEvent(cmdletName, localProps);
+
+            // Create the logging entry
+            _telemetryClient.TrackEvent(cmdletName, localProps);
         }
     }
 }
